% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.separability.R
\name{classes.separability}
\alias{classes.separability}
\title{Visual tests for classes separability}
\usage{
classes.separability(data, target_classes = NULL, scatterplot = FALSE)
}
\arguments{
\item{data}{a dataframe with reflectance data containing pixel values for Sentinel optical bands
B2, B3, B4, B5, B6, B7, B8, B8A, B11, B12}

\item{target_classes}{a list of any number of classes of surface to analyse.
All classes available in data are taken by default.}

\item{scatterplot}{logical. If TRUE, band-to-band scatter plots will be written
to the working directory. FALSE by default.}
}
\value{
list of four plots: three PCA for different combination of PCs, a screeplot
to PCA diagnostics.
}
\description{
Classes separability is a keystone of supervised image classification.
To evaluate this, {sprectralR} provides Principal Component Analysis (PCA) for
reflectance data, showing the variability, separability, and compactness
of defined classes. Well separated, compact clouds of points reflect optically
distinguished classes. Whereas spread, split, or overlapped clouds of points
indicate that user's classes are possibly intermixed or, maybe, just cannot be
separated with the present remote sensing data. User can also make and write to
the working directory a series of scatter plots for each possible band combination,
to find out in which band classes are intermixed.
}
\examples{
# Load example data
load(system.file("testdata/reflectance_test_data.RData", package = "spectralR"))

# Check available class names
levels(reflectance$label)

# Explore separability in all classes
separability_all <- classes.separability(reflectance)

# Plot data values in a gradient of two main Principal Components
separability_all

# Another way to plot Principal Components 1 and 2
separability_all$PC1.PC2

# Plot PC2 and PC3
separability_all$PC2.PC3

# Scree plot (Fraction of total variance covered by each Principal Component)
separability_all$screeplot

# Explore separability for three selected classes, and write a pand-to-band plots
# into the working directory
separability_some_classes <- classes.separability(data = reflectance,
                                     target_classes = list("cropland", "meadow", "reed"),
                                     scatterplot = TRUE)

separability_some_classes

}
