% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralR.R
\name{prepare.vector.data}
\alias{prepare.vector.data}
\title{prepare.vector.data:}
\usage{
prepare.vector.data(shapefile_name, label_field)
}
\arguments{
\item{shapefile_name}{shapefile name (should lie within working directory, using absolute
paths were not tested)}

\item{label_field}{name of the field which contains class labels}
}
\value{
Result of calculation
}
\description{
Function takes shapefile with polygons of different classes of surface (habitats, crops,
vegetation, etc.), and retrieves ready-for-sampling sf object.
}
