% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.tests.R
\name{stat.tests}
\alias{stat.tests}
\title{Statistical tests for two surface classes}
\usage{
stat.tests(data, target_classes)
}
\arguments{
\item{data}{reflectance data as dataframe with pixel values for Sentinel optical bands
B2, B3, B4, B5, B6, B7, B8, B8A, B11, B12}

\item{target_classes}{list of two classes of surface to analyse.}
}
\value{
list of four lists: .$normality, .$homogeneity, .$imbalance, .$plot
}
\description{
Exploratory analysis of the pair of user-specified classes.
Includes two statistical tests - Shapiro-Wilk normality test and Bartlett
test of homogeneity of variances, - as well as class imbalance assessment
and histograms of distribution of reflectance values per band.
}
\examples{
# Load example data
load(system.file("testdata/reflectance_test_data.RData", package = "spectralR"))

# Check available class names
levels(reflectance$label)

# Calculate normality, homogeneity, class imbalance, and plot the distribution
stattests <- stat.tests(data = reflectance,
                        target_classes = list("deciduous_forest", "coniferous_forest"))

stattests$normality # to look at Shapiro-Wilk test for normality

stattests$homogeneity # to look at Bartlett Test of Homogeneity of Variances

stattests$imbalance # to explore possible class imbalance

stattests$plot # to look at reflectance values distribution per band

}
