% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin.plot.R
\name{violin.plot}
\alias{violin.plot}
\title{Reflectance per bands as violin plots for each surface class}
\usage{
violin.plot(data)
}
\arguments{
\item{data}{reflectance data as dataframe with pixel values for Sentinel optical bands
B2, B3, B4, B5, B6, B7, B8, B8A, B11, B12}
}
\value{
ggplot2 object with basic visual aesthetics. Default aesthetics is
violin plot for each satellite band (geom_violin).
See https://ggplot2.tidyverse.org/reference/geom_violin.html for more details.
}
\description{
Reflectance per bands as violin plots for each surface class
}
\examples{
p3 <- violin.plot(
  data = reflectance
  )

p3

p3 +
  labs(x='Surface class',y='Reflectance',
      fill="Surface classes",
      title = "Reflectance for different surface classes",
      caption='Data: Sentinel-2 Level-2A')+
   theme_minimal()
}
