% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within.band.analysis.R
\name{within.band.analysis}
\alias{within.band.analysis}
\title{Per-band visual exploration of spectral profiles of two user-specified classes}
\usage{
\method{within}{band.analysis}(data, target_classes, method = list("wilcoxon", "t-test"))
}
\arguments{
\item{data}{reflectance data as dataframe with pixel values for Sentinel optical bands
B2, B3, B4, B5, B6, B7, B8, B8A, B11, B12}

\item{target_classes}{list of two classes of surface to analyse.}

\item{method}{available statistical method for significance in differences
("wilcoxon", or "t-test")}
}
\value{
list of four: .$boxplot, .$bandplot, .$confidenceIntervals, .$significantDifferences
}
\description{
Visual per-band analysis of the pair of user-specified classes.
Returns two types of plots, each accompanied with information on statistical differences
in reflectance values in pair of classes per band. Three methods for exploring
differences are used: (i) either parametric Student's t-Test or non-parametric
Wilcoxon test, (ii) level of differences in class means, and (iii) confidence intervals
for a differences in class means.
}
\examples{
# Load example data
load(system.file("testdata/reflectance_test_data.RData", package = "spectralR"))

# Check available class names
levels(reflectance$label)

# Analyze differences in spectral profiles of selected classes per band
two.class.difs <- within.band.analysis(data = reflectance,
                                       target_classes = list("deciduous_forest", "coniferous_forest"),
                                       method = "t-test")
# Boxplot across all bands
two.class.difs$boxplot

# Statistical summary plot across bands
two.class.difs$bandplot

# Differences in means, lower, upper, and full confidence intervals for a differences in class means.
two.class.difs$confidenceIntervals

# Significance in differences between selected classes per band, according to selected statistical test.
two.class.difs$significantDifferences

}
