% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pixel.data.R
\name{get.pixel.data}
\alias{get.pixel.data}
\title{Obtain Sentinel-2 spectral reflectance data for user-defined vector polygons}
\usage{
get.pixel.data(sf_data, startday, endday, cloud_threshold, scale_value)
}
\arguments{
\item{sf_data}{polygons of surface classes as a sf object, prepared on previous step}

\item{startday}{starting day for Sentinel image collection, as "YYYY-MM-DD". See Note 1 below}

\item{endday}{final day for Sentinel image collection, as "YYYY-MM-DD"}

\item{cloud_threshold}{maximum per cent of cloud-covered pixels per image by which individual}

\item{scale_value}{scale of resulting satellite images in meters (pixel size). See Note 2 below}
}
\value{
A dataframe (non-spatial) with unscaled reflectance data for each pixel of median satellite
image, for each optical band of Sentinel-2 sensor, marked according to the label of surface class from
the polygons.
}
\description{
Function takes sf polygon object, obtained in the previous step, and retrieves data
frame with brightness values for each pixel intersected with polygons, for each
optical band of Sentinel-2 sensor, marked according to the label of surface class from
the polygons.
}
\examples{
reflectance = get.pixel.data(sf_df, "2019-05-15", "2019-06-30", 10, 100)
}
